+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-12-11
+
+### Added
+- New diagnostic sensor "Operation State" showing detailed heat pump operation mode
+- New diagnostic sensor "Compressor Cycle Time" measuring average time between compressor starts
+
+### Changed
+- Updated French and English translations for new sensors
+
 ## [1.3.4] - 2024-12-05
 
 ### Changed